moon_server/requirements.txt
# Moon Server Requirements
Django>=1.4
django_compressor>=1.3
django_openstack_auth>=1.1.4
simplejson
jsonpatch
jsonpointer
python-keystoneclient
python-novaclient
